+++ changelog.md
@@ -1,3 +1,9 @@
+5.0.0
+  - breaking change: moved comment and clientId arguments from MetaApiConnection trade methods to options argument
+  - added magic trade option to let you specify distinct magic number (expert advisor id) on each trade
+  - added manualTrades field to account model so that it is possible to configure if MetaApi should place manual trades on the account
+  - prepare MetatraderAccountApi class for upcoming breaking change in the API
+  
 4.0.0
   - add fields to trade result to match upcoming MetaApi contract
   - breaking change: throw TradeException in case of trade error